third_party/WebKit/LayoutTests/svg/animations/repeatn-remove-add-animation-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests


This removes and adds an animation element while the animation is repeating

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 255
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 255
PASS colorComponents[3] is 0
PASS successfullyParsed is true

TEST COMPLETE
